Split apart grammar.py
