+++ test/core/parser/grammar_test.py
@@ -5,12 +5,12 @@ import logging
 
 from sqlfluff.core.parser import RootParseContext, KeywordSegment
 from sqlfluff.core.parser.segments_common import EphemeralSegment
+from sqlfluff.core.parser.grammar.base import BaseGrammar
 from sqlfluff.core.parser.grammar import (
     OneOf,
     Sequence,
     GreedyUntil,
     Delimited,
-    BaseGrammar,
     StartsWith,
     Anything,
     Nothing,
